requirements.txt
discord.py
google-cloud-dialogflow
python-dotenv
flask